+++ spi-impl/series/series-dao/src/main/java/org/n52/io/extension/metadata/DatabaseMetadataExtension.java
@@ -28,23 +28,20 @@
  */
 package org.n52.io.extension.metadata;
 
-import java.util.Collections;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Set;
 
-import org.hibernate.Session;
 import org.n52.io.request.IoParameters;
 import org.n52.io.response.ParameterOutput;
 import org.n52.io.response.extension.MetadataExtension;
-import org.n52.series.db.da.SessionAwareRepository;
+import org.springframework.beans.factory.annotation.Autowired;
 
 public class DatabaseMetadataExtension extends MetadataExtension<ParameterOutput> {
 
     private static final String EXTENSION_NAME = "databaseMetadata";
 
-    private final MetadataRepository repository = new MetadataRepository();
+    @Autowired
+    private MetadataRepository repository;
 
     @Override
     public String getExtensionName() {
@@ -62,44 +59,6 @@ public class DatabaseMetadataExtension extends MetadataExtension<ParameterOutput
         for (String fieldName : fieldNames) {
             output.addExtra(fieldName);
         }
-    }
-
-    private class MetadataRepository extends SessionAwareRepository {
-
-        private List<String> getFieldNames(String id) {
-            Session session = getSession();
-            try {
-                DatabaseMetadataDao dao = new DatabaseMetadataDao(session);
-                return dao.getMetadataNames(parseId(id));
-            } finally {
-                returnSession(session);
-            }
-        }
-
-        private Map<String, Object> getExtras(ParameterOutput output, IoParameters parameters) {
-            Session session = getSession();
-            try {
-                DatabaseMetadataDao dao = new DatabaseMetadataDao(session);
-                final Set<String> fields = parameters.getFields();
-                return fields == null
-                        ? convertToOutputs(dao.getAllFor(parseId(output.getId())))
-                        : convertToOutputs(dao.getSelected(parseId(output.getId()), fields));
-            } finally {
-                returnSession(session);
-            }
-        }
-
-        private Map<String, Object> convertToOutputs(List<MetadataEntity<?>> allInstances) {
-            if (allInstances == null) {
-                return Collections.emptyMap();
-            }
-            Map<String, Object> outputs = new HashMap<>();
-            for (MetadataEntity<?> entity : allInstances) {
-                outputs.put(entity.getName(), entity.toOutput());
-            }
-            return outputs;
-        }
-
     }
 
 }
